Remove index files from action-links, accounts and assets
