--- app/accounts/index.js
@@ -1,1 +1,0 @@
-export * from './server/index';
