+++ app/action-links/client/index.js
@@ -1,5 +1,4 @@
-import { actionLinks } from '../both/lib/actionLinks';
-import './lib/actionLinks';
+import { actionLinks } from './lib/actionLinks';
 import './init';
 import './stylesheets/actionLinks.css';
