+++ app/action-links/client/init.js
@@ -1,10 +1,10 @@
 import { Blaze } from 'meteor/blaze';
 import { Template } from 'meteor/templating';
 
-import { handleError } from '../../utils';
-import { fireGlobalEvent, Layout } from '../../ui-utils';
+import { handleError } from '../../utils/client';
+import { fireGlobalEvent, Layout } from '../../ui-utils/client';
 import { messageArgs } from '../../ui-utils/client/lib/messageArgs';
-import { actionLinks } from '../both/lib/actionLinks';
+import { actionLinks } from './lib/actionLinks';
 
 
 Template.room.events({
