--- app/action-links/index.js
@@ -1,8 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-
-if (Meteor.isClient) {
-	module.exports = require('./client/index.js');
-}
-if (Meteor.isServer) {
-	module.exports = require('./server/index.js');
-}
