+++ app/action-links/server/actionLinkHandler.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 
-import { actionLinks } from '../both/lib/actionLinks';
+import { actionLinks } from './lib/actionLinks';
 // Action Links Handler. This method will be called off the client.
 
 Meteor.methods({
