+++ app/action-links/server/index.js
@@ -1,4 +1,4 @@
-import { actionLinks } from '../both/lib/actionLinks';
+import { actionLinks } from './lib/actionLinks';
 import './actionLinkHandler';
 
 export {
