+++ app/blockstack/server/routes.js
@@ -2,7 +2,7 @@ import { Meteor } from 'meteor/meteor';
 import { WebApp } from 'meteor/webapp';
 
 import { settings } from '../../settings';
-import { RocketChatAssets } from '../../assets';
+import { RocketChatAssets } from '../../assets/server';
 
 WebApp.connectHandlers.use('/_blockstack/manifest', Meteor.bindEnvironment(function(req, res) {
 	const name = settings.get('Site_Name');
