+++ app/livechat/client/lib/messageTypes.js
@@ -1,0 +1,5 @@
+import { actionLinks } from '../../../action-links/client';
+
+actionLinks.register('createLivechatCall', function(message, params, instance) {
+	instance.tabBar.open('video');
+});
