+++ app/videobridge/client/actionLink.js
@@ -2,7 +2,7 @@ import { Session } from 'meteor/session';
 import { TAPi18n } from 'meteor/rocketchat:tap-i18n';
 import toastr from 'toastr';
 
-import { actionLinks } from '../../action-links';
+import { actionLinks } from '../../action-links/client';
 import { Rooms } from '../../models';
 
 actionLinks.register('joinJitsiCall', function(message, params, instance) {
