+++ app/videobridge/server/actionLink.js
@@ -1,4 +1,4 @@
-import { actionLinks } from '../../action-links';
+import { actionLinks } from '../../action-links/server';
 
 actionLinks.register('joinJitsiCall', function(/* message, params*/) {
