+++ server/importPackages.js
@@ -1,10 +1,10 @@
 import '../app/cors/server';
 import '../app/sms';
 import '../app/2fa/server';
-import '../app/accounts';
+import '../app/accounts/server';
 import '../app/analytics/server';
 import '../app/api';
-import '../app/assets';
+import '../app/assets/server';
 import '../app/authorization';
 import '../app/autolinker/server';
 import '../app/autotranslate/server';
@@ -109,6 +109,6 @@ import '../app/callbacks';
 import '../app/notifications';
 import '../app/promises/server';
 import '../app/ui-utils';
-import '../app/action-links';
+import '../app/action-links/server';
 import '../app/reactions/server';
 import '../app/livechat/server';
